GH-1748 expose overnight indexed coupon pricer
